+++ packages/engine/src/component/CustomSchemaForm/components/Setters/ActionFlowSetter/node/RequestAPINode/index.tsx
@@ -22,12 +22,22 @@ import {
 import styles from './style.module.scss';
 import { useActionFlow } from '../../context';
 import { CFiledWithSwitchSetter } from '@/component/CustomSchemaForm/components/CFiledWithSwitchSetter';
+import { RightPanelConfig } from '@/plugins/RightPanel/type';
 
 export type TRequestAPINode = Node<TLogicRequestAPIItem, 'RequestAPINode'>;
 
 export const RequestAPINode = (props: NodeProps<TRequestAPINode>) => {
   const { data, isConnectable } = props;
-  const { onDataChange } = useActionFlow();
+  const { onDataChange, pluginCtx } = useActionFlow();
+  const pluginConfig: RightPanelConfig = pluginCtx.config || {};
+
+  const CustomAPISelectInput: any = useMemo(() => {
+    if (pluginConfig.requestAPINode?.customAPIInput) {
+      return pluginConfig.requestAPINode?.customAPIInput;
+    }
+    return Input;
+  }, [pluginConfig.requestAPINode?.customAPIInput]);
+
   const reactFlowInstance = useReactFlow();
   ensureKeyExist(data, DEV_CONFIG_KEY, {});
   const devConfigObj = data[DEV_CONFIG_KEY]!;
@@ -60,6 +70,7 @@ export const RequestAPINode = (props: NodeProps<TRequestAPINode>) => {
     } as TLogicRequestAPIItem;
     formRef.current?.setFields(newVal);
     setFormValue(newVal);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   const updateKeySetterConfig = (keyPaths: string[], setterName: string) => {
@@ -172,6 +183,11 @@ export const RequestAPINode = (props: NodeProps<TRequestAPINode>) => {
             <>
               <InputHandle type="target" isConnectable={isConnectable} />
               <CreateNewNodePopup
+                style={{
+                  position: 'absolute',
+                  left: '50%',
+                  bottom: '0',
+                }}
                 title="请求成功时"
                 onNewNodeAdd={(data) => handleNewNodeAdd(data, OUTPUT_HANDLE_ID)}
                 disabled={isOutputHandleConnected}
@@ -191,14 +207,15 @@ export const RequestAPINode = (props: NodeProps<TRequestAPINode>) => {
                 disabled={isAfterFailedResponseHandleConnected}
               >
                 <OutputHandle
-                  id={REQUEST_API_FAILED_HANDLE_ID}
-                  isConnectable={isConnectable}
                   style={{
+                    left: '75%',
+                    bottom: '0',
                     width: '10px',
                     height: '10px',
                     background: '#ff4d4f',
-                    left: '75%',
                   }}
+                  id={REQUEST_API_FAILED_HANDLE_ID}
+                  isConnectable={isConnectable}
                 />
               </CreateNewNodePopup>
             </>
@@ -228,7 +245,7 @@ export const RequestAPINode = (props: NodeProps<TRequestAPINode>) => {
                   valueChangeEventName="onChange"
                   formatEventValue={(el) => el.target.value}
                 >
-                  <Input />
+                  <CustomAPISelectInput form={formRef} />
                 </CField>
               </div>
               <div className={styles.line}>