+++ packages/engine/src/index.tsx
@@ -188,3 +188,5 @@ export * from './utils/index';
 export { registerCustomSetter } from './component/CustomSchemaForm/components/Form';
 
 export * from './type';
+export * from './core/pluginManager';
+export * from './core/assetPackagesListManage';